# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "salus documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "appointments"
                ],
                "summary": "Get all appointments for the authenticated user",
                "operationId": "b8c7b364f8969997096bb5cc471278c2",
                "responses": {
                    "200": {
                        "description": "List of user appointments"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointments not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "appointments"
                ],
                "summary": "Update appointment",
                "description": "Update the specified resource in storage.",
                "operationId": "b7b569d19bf3887d68d3ab5f9a3dfddc",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_date",
                                    "doctor_id"
                                ],
                                "properties": {
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending or confirmed",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ]
                                    },
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "appointment Updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Store a newly created resource in storage.",
                "operationId": "a115e378e7f96db9e8a6537cbdd95921",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_date",
                                    "doctor_id"
                                ],
                                "properties": {
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-03-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending or confirmed",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ]
                                    },
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symptom created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "get": {
                "tags": [
                    "appointments"
                ],
                "summary": "Get one appointments for the authenticated user",
                "description": "Display the specified resource.",
                "operationId": "a38fbb7f74ce31ce3a372d53c082515b",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user appointment"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "appointments"
                ],
                "summary": "Delete appointments for the authenticated user",
                "description": "Remove the specified resource from storage.",
                "operationId": "340a1d1e4b71864ef4d72553cf13c68a",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user appointment"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "appointment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new account",
                "operationId": "c758f54bda9079f9ebe361693213d40c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "chahd"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "chahd@gmail.com"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "user created"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "logged in",
                "operationId": "1da689c688cc418e88f9e9d668f01b16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "chahd@gmail.com"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "user created"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "logged out",
                "operationId": "e8863ea5beee6c0b123d63b0aba831b0",
                "responses": {
                    "201": {
                        "description": "user created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sunctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}